[REFACTOR] Simplify lib exports
In the past, we had to import/export some classes because they were defined as
default in their source files.

The following classes are no more exported as default:
  - BpmnVisualization
  - IconPainter
  - ShapeUtil
  - StyleConfigurator

Also rename files that with no more default export.
