+++ src/component/mxgraph/shape/render/index.ts
@@ -14,5 +14,5 @@
  * limitations under the License.
  */
 export * from './BpmnCanvas';
-export * from './IconPainter';
+export * from './icon-painter';
 export * from './render-types';
